pyproject.toml
[tool.poetry]
name = "reddit-presenter"
version = "0.1.0"
description = ""
authors = ["Mads Kongstad Laugesen <kongstad.mads@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.9"
praw = "^7.5.0"
pandas = "^1.5.2"
python-pptx = "^0.6.21"
numpy = "^1.23.5"
matplotlib = "^3.6.2"
seaborn = "^0.12.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
